Recreate events commands for Atlas
