+++ internal/cli/atlas/events/orgs_list.go
@@ -1,10 +1,10 @@
-// Copyright 2021 MongoDB Inc
+// Copyright 2023 MongoDB Inc
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//      http://www.apache.org/licenses/LICENSE-2.0
+//	http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -28,14 +28,14 @@ import (
 	atlas "go.mongodb.org/atlas/mongodbatlas"
 )
 
-type projectListOpts struct {
+type orgListOpts struct {
+	cli.OutputOpts
 	EventListOpts
 	cli.GlobalOpts
-	cli.OutputOpts
-	store store.ProjectEventLister
+	store store.OrganizationEventLister
 }
 
-func (opts *projectListOpts) initStore(ctx context.Context) func() error {
+func (opts *orgListOpts) initStore(ctx context.Context) func() error {
 	return func() error {
 		var err error
 		opts.store, err = store.New(store.AuthenticatedPreset(config.Default()), store.WithContext(ctx))
@@ -43,13 +43,12 @@ func (opts *projectListOpts) initStore(ctx context.Context) func() error {
 	}
 }
 
-func (opts *projectListOpts) Run() error {
+func (opts *orgListOpts) Run() error {
 	listOpts := opts.newEventListOptions()
 
 	var r *atlas.EventResponse
 	var err error
-	r, err = opts.store.ProjectEvents(opts.ConfigProjectID(), listOpts)
-
+	r, err = opts.store.OrganizationEvents(opts.ConfigOrgID(), listOpts)
 	if err != nil {
 		return err
 	}
@@ -57,26 +56,27 @@ func (opts *projectListOpts) Run() error {
 	return opts.Print(r)
 }
 
-// ProjectListBuilder
+// OrgListBuilder
 //
-//	mongocli atlas event(s) list
+//	atlas event(s) list
 //
+// [--orgId orgId]
 // [--page N]
 // [--limit N]
 // [--minDate minDate]
 // [--maxDate maxDate].
-func ProjectListBuilder() *cobra.Command {
-	opts := &projectListOpts{}
+func OrgListBuilder() *cobra.Command {
+	opts := &orgListOpts{}
 	cmd := &cobra.Command{
 		Use:     "list",
-		Short:   "Return all events for the specified project.",
+		Short:   "Return all events for the specified organization.",
 		Aliases: []string{"ls"},
 		Args:    require.NoArgs,
-		Example: fmt.Sprintf(`  # Return a JSON-formatted list of events for the project with the ID 5e2211c17a3e5a48f5497de3:
-  %s events projects list --Id 5e2211c17a3e5a48f5497de3 --output json`, cli.ExampleAtlasEntryPoint()),
+		Example: fmt.Sprintf(`  # Return a JSON-formatted list of events for the organization with the ID 5dd5a6b6f10fab1d71a58495:
+  %s events organizations list --orgId 5dd5a6b6f10fab1d71a58495 --output json`, cli.ExampleAtlasEntryPoint()),
 		PreRunE: func(cmd *cobra.Command, args []string) error {
 			return opts.PreRunE(
-				opts.ValidateProjectID,
+				opts.ValidateOrgID,
 				opts.initStore(cmd.Context()),
 				opts.InitOutput(cmd.OutOrStdout(), listTemplate),
 			)
@@ -93,24 +93,23 @@ func ProjectListBuilder() *cobra.Command {
 	cmd.Flags().StringVar(&opts.MaxDate, flag.MaxDate, "", usage.MaxDate)
 	cmd.Flags().StringVar(&opts.MinDate, flag.MinDate, "", usage.MinDate)
 
-	cmd.Flags().StringVar(&opts.ProjectID, flag.ProjectID, "", usage.ProjectID)
+	cmd.Flags().StringVar(&opts.OrgID, flag.OrgID, "", usage.OrgID)
 	cmd.Flags().StringVarP(&opts.Output, flag.Output, flag.OutputShort, "", usage.FormatOut)
 	_ = cmd.RegisterFlagCompletionFunc(flag.Output, opts.AutoCompleteOutputFlag())
 
 	return cmd
 }
 
-func ProjectsBuilder() *cobra.Command {
-	const use = "projects"
+func OrgsBuilder() *cobra.Command {
+	const use = "organizations"
 	cmd := &cobra.Command{
 		Use:     use,
-		Short:   "Project operations.",
-		Long:    "List projects events.",
-		Aliases: cli.GenerateAliases(use),
+		Short:   "Organization operations.",
+		Long:    "List organization events.",
+		Aliases: cli.GenerateAliases(use, "orgs", "org"),
 	}
 	cmd.AddCommand(
-		ProjectListBuilder(),
+		OrgListBuilder(),
 	)
-
 	return cmd
 }
