+++ internal/cli/cloudmanager/cloud_manager.go
@@ -16,8 +16,8 @@ package cloudmanager
 
 import (
 	"github.com/mongodb/mongodb-atlas-cli/internal/cli"
-	"github.com/mongodb/mongodb-atlas-cli/internal/cli/events"
 	"github.com/mongodb/mongodb-atlas-cli/internal/cli/mongocli/alerts"
+	"github.com/mongodb/mongodb-atlas-cli/internal/cli/mongocli/events"
 	"github.com/mongodb/mongodb-atlas-cli/internal/cli/mongocli/performanceadvisor"
 	"github.com/mongodb/mongodb-atlas-cli/internal/cli/opsmanager/agents"
 	"github.com/mongodb/mongodb-atlas-cli/internal/cli/opsmanager/automation"
