+++ internal/cli/mongocli/events/orgs_list_test.go
@@ -1,0 +1,71 @@
+// Copyright 2023 MongoDB Inc
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package events
+
+import (
+	"testing"
+
+	"github.com/golang/mock/gomock"
+	"github.com/mongodb/mongodb-atlas-cli/internal/flag"
+	"github.com/mongodb/mongodb-atlas-cli/internal/mocks"
+	"github.com/mongodb/mongodb-atlas-cli/internal/test"
+	"go.mongodb.org/atlas/mongodbatlas"
+)
+
+func Test_orgListOpts_Run(t *testing.T) {
+	ctrl := gomock.NewController(t)
+	mockStore := mocks.NewMockOrganizationEventLister(ctrl)
+
+	expected := &mongodbatlas.EventResponse{}
+	listOpts := &orgListOpts{
+		store: mockStore,
+	}
+	listOpts.OrgID = "1"
+
+	mockStore.
+		EXPECT().OrganizationEvents(listOpts.OrgID, listOpts.newEventListOptions()).
+		Return(expected, nil).
+		Times(1)
+
+	if err := listOpts.Run(); err != nil {
+		t.Fatalf("Run() unexpected error: %v", err)
+	}
+}
+
+func TestOrgListBuilder(t *testing.T) {
+	test.CmdValidator(
+		t,
+		OrgListBuilder(),
+		0,
+		[]string{
+			flag.Limit,
+			flag.Page,
+			flag.Output,
+			flag.OrgID,
+			flag.TypeFlag,
+			flag.MaxDate,
+			flag.MinDate,
+		},
+	)
+}
+
+func TestOrgsBuilder(t *testing.T) {
+	test.CmdValidator(
+		t,
+		OrgsBuilder(),
+		1,
+		[]string{},
+	)
+}
