+++ internal/cli/root/atlas/builder.go
@@ -36,6 +36,7 @@ import (
 	"github.com/mongodb/mongodb-atlas-cli/internal/cli/atlas/customdns"
 	"github.com/mongodb/mongodb-atlas-cli/internal/cli/atlas/datalake"
 	"github.com/mongodb/mongodb-atlas-cli/internal/cli/atlas/dbusers"
+	"github.com/mongodb/mongodb-atlas-cli/internal/cli/atlas/events"
 	"github.com/mongodb/mongodb-atlas-cli/internal/cli/atlas/integrations"
 	"github.com/mongodb/mongodb-atlas-cli/internal/cli/atlas/kubernetes"
 	"github.com/mongodb/mongodb-atlas-cli/internal/cli/atlas/livemigrations"
@@ -55,7 +56,6 @@ import (
 	"github.com/mongodb/mongodb-atlas-cli/internal/cli/atlas/teams"
 	"github.com/mongodb/mongodb-atlas-cli/internal/cli/atlas/users"
 	"github.com/mongodb/mongodb-atlas-cli/internal/cli/auth"
-	"github.com/mongodb/mongodb-atlas-cli/internal/cli/events"
 	"github.com/mongodb/mongodb-atlas-cli/internal/cli/figautocomplete"
 	"github.com/mongodb/mongodb-atlas-cli/internal/config"
 	"github.com/mongodb/mongodb-atlas-cli/internal/flag"
